updated cpus function, stilldoesnt work
