+++ test.c
@@ -1,0 +1,17 @@
+#define _GNU_SOURCE
+#include <stdio.h>
+#include <stdlib.h>
+
+int main(int argc, char **argv)
+{
+   FILE *cpuinfo = fopen("/proc/cpuinfo", "rb");
+   char *arg = 0;
+   size_t size = 0;
+   while(getdelim(&arg, &size, 0, cpuinfo) != -1)
+   {
+      puts(arg);
+   }
+   free(arg);
+   fclose(cpuinfo);
+   return 0;
+}
